mapesolutions/web/requirements.txt
dj-database-url==0.5.0
Django<1.12
django-filter==1.1.0
djangorestframework==3.9.1
djangorestframework-filters==0.10.2.post0
djangorestframework-guardian==0.1.1
psycopg2-binary==2.7.5
simplejson==3.16.0


